optimized relational model, make api for rent logs (need verification, canceled, return), optimized logic)

+++ app/Http/Controllers/BookController.php
@@ -3,10 +3,11 @@
 namespace App\Http\Controllers;
 
 use App\Models\Book;
+use App\Models\Favorite;
 use App\Models\Type;
 use Exception;
-use Illuminate\Auth\Events\Validated;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Facades\Validator;
 use Illuminate\Support\Str;
@@ -28,10 +29,17 @@ class BookController extends Controller
     {
         $book = Book::findOrFail($slug);
 
-        return response()->json([
-            'status' => 'success',
-            'data' => $book
-        ]);
+        if ($book) {
+            return response()->json([
+                'status' => 'success',
+                'data' => $book
+            ]);
+        } else {
+            return response()->json([
+                'status' => 'error',
+                'message' => 'BOOK NOT FOUND'
+            ]);
+        }
     }
 
     public function add(Request $req)
@@ -101,55 +109,69 @@ class BookController extends Controller
     {
         try {
             $book = Book::where('slug' . $slug)->first();
-            $validator = Validator::make($req->all(), [
-                'type_id' => 'required',
-                'total_book' => 'required',
-                'title' => 'required|unique:books,title,'.$book->id,
-                'writer' => 'required',
-                'publisher' => '',
-                'description' => 'required',
-                'year' => '',
-                'page' => 'required',
-                'cover' => '',
-            ]);
+            if ($book) {
+                $validator = Validator::make($req->all(), [
+                    'type_id' => 'required',
+                    'total_book' => 'required',
+                    'title' => 'required|unique:books,title,' . $book->id,
+                    'writer' => 'required',
+                    'publisher' => '',
+                    'description' => 'required',
+                    'year' => '',
+                    'page' => 'required',
+                    'cover' => '',
+                ]);
 
-            $data = [
-                'type_id' => $req->type_id,
-                'total_book' => $req->total_book,
-                'title' => $req->title,
-                'writer' => $req->writer,
-                'publisher' => $req->publisher,
-                'description' => $req->description,
-                'year' => $req->year,
-                'page' => $req->page,
-                'cover' => $book->cover,
-            ];
-
-            if ($req->hasFile('cover')) {
-                if ($data['cover']) {
-                    Storage::delete($data['cover']);
+                if ($validator->fails()) {
+                    return response()->json([
+                        'status' => 'error',
+                        'message' => $validator->errors()
+                    ]);
                 }
+
+                $data = [
+                    'type_id' => $req->type_id,
+                    'total_book' => $req->total_book,
+                    'title' => $req->title,
+                    'writer' => $req->writer,
+                    'publisher' => $req->publisher,
+                    'description' => $req->description,
+                    'year' => $req->year,
+                    'page' => $req->page,
+                    'cover' => $book->cover,
+                ];
+
+                if ($req->hasFile('cover')) {
+                    if ($data['cover']) {
+                        Storage::delete($data['cover']);
+                    }
+                } else {
+                    $fileName = Str::slug($req->title) . '.' . $req->file('cover')->getClientOriginalExtension();
+                    $path = $req->file('cover')->storeAs('books', $fileName);
+                    $data['cover'] = '/storage/' . $path;
+                }
+
+                $book->slug = null;
+                $book->save($data);
+
+                if ($req->has('categories')) {
+                    $book->categories()->sync($req->input('categories'));
+                } else {
+                    $book->categories()->detach();
+                }
+
+
+                return response()->json([
+                    'status' => 'success',
+                    'message' => 'Successfully Updated Book Information',
+                    'data' => $book,
+                ]);
             } else {
-                $fileName = Str::slug($req->title) . '.' . $req->file('cover')->getClientOriginalExtension();
-                $path = $req->file('cover')->storeAs('books', $fileName);
-                $data['cover'] = '/storage/' . $path;
+                return response()->json([
+                    'status' => 'error',
+                    'message' => 'BOOK NOT FOUND'
+                ]);
             }
-
-            $book->save($data);
-
-            if ($req->has('categories')) {
-                $book->categories()->sync($req->input('categories'));
-            } else {
-                $book->categories()->detach();
-            }
-
-
-            return response()->json([
-                'status' => 'success',
-                'message' => 'Successfully Updated Book Information',
-                'data' => $book,
-            ]);
-
         } catch (Exception $e) {
             return response()->json([
                 'status' => 'error',
@@ -165,19 +187,26 @@ class BookController extends Controller
             // RENTLOGS, REVIEW, TYPE, CATEGORIES, FAVORITES
             $book = Book::where('slug', $slug)->first();
 
-            $book->types->delete();
-            // $book->rentlogs()->delete();
-            // $book->reviews()->delete();
-            // $book->favorites()->delete();
+            if ($book) {
+                $book->types->delete();
+                $book->rentlogs()->delete();
+                $book->reviews()->delete();
+                $book->favorites()->delete();
+                $book->categories()->detach();
 
-            Storage::delete($book->cover);
-            $book->categories()->detach();
-            $book->delete();
+                Storage::delete($book->cover);
+                $book->delete();
 
-            return response()->json([
-                'status' => 'success',
-                'message' => 'Successfully Deleted A Book',
-            ]);
+                return response()->json([
+                    'status' => 'success',
+                    'message' => 'Successfully Deleted A Book',
+                ]);
+            } else {
+                return response()->json([
+                    'status' => 'error',
+                    'message' => 'BOOK NOT FOUND'
+                ]);
+            }
         } catch (Exception $e) {
             return response()->json([
                 'status' => 'error',
@@ -187,4 +216,42 @@ class BookController extends Controller
     }
 
 
+
+
+    // END CRUD || END CRUD || END CRUD || END CRUD || END CRUD || END CRUD || END CRUD || END CRUD || END CRUD ||
+
+
+    public function fav(Request $req, $code)
+    {
+        $fav = Favorite::where('book_id', $req->book_id)->where('user_id', Auth::user()->id)->first();
+
+        try {
+            if ($fav) {
+                $data = [
+                    'user_id' => Auth::user(),
+                    'code' => Str::random(4) . '-' . Auth::user()->username,
+                    'book_id' => $req->book_id
+                ];
+
+                Favorite::create($data);
+
+                return response()->json([
+                    'status' => 'success',
+                    'message' => 'SUCCESSFULLY FAVORITE THIS BOOK',
+                ]);
+            } else {
+                $fav->delete();
+
+                return response()->json([
+                    'status' => 'success',
+                    'message' => 'SUCCESSFULLY UNFAVORITE THIS BOOK',
+                ]);
+            }
+        } catch (Exception $e) {
+            return response()->json([
+                'status' => 'error',
+                'message' => $e->getMessage()
+            ]);
+        }
+    }
 }
